Upgrade dependencies + fixes for breaking changes
