+++ .dart_tool/package_config_subset
@@ -1,5 +1,5 @@
 calculadora_imc
-2.1
+2.12
 file:///home/edilson/GitHub/repo/public/calculadora_imc/
 file:///home/edilson/GitHub/repo/public/calculadora_imc/lib/
 async
