+++ .packages
@@ -3,7 +3,7 @@
 # 
 # For more info see: https://dart.dev/go/dot-packages-deprecation
 # 
-# Generated by pub on 2022-01-17 15:02:34.876031.
+# Generated by pub on 2022-01-17 15:07:18.797781.
 async:file:///home/edilson/development/flutter/.pub-cache/hosted/pub.dartlang.org/async-2.8.2/lib/
 boolean_selector:file:///home/edilson/development/flutter/.pub-cache/hosted/pub.dartlang.org/boolean_selector-2.1.0/lib/
 characters:file:///home/edilson/development/flutter/.pub-cache/hosted/pub.dartlang.org/characters-1.2.0/lib/
