+++ lib/main.dart
@@ -1,164 +1,12 @@
 import 'package:flutter/material.dart';
 
+import 'screens/home_screen.dart';
+
 void main() {
   runApp(
     MaterialApp(
-      home: Home(),
+      home: const HomeScreen(),
       debugShowCheckedModeBanner: false,
     ),
   );
-}
-
-class Home extends StatefulWidget {
-  @override
-  _HomeState createState() => _HomeState();
-}
-
-class _HomeState extends State<Home> {
-  TextEditingController weightController = TextEditingController();
-  TextEditingController heightController = TextEditingController();
-
-  GlobalKey<FormState> _formKey = GlobalKey<FormState>();
-
-  String _infoText = 'Digite os seus dados!';
-
-  void _resetFields() {
-    weightController.text = '';
-    heightController.text = '';
-    setState(() {
-      _infoText = 'Digite os seus dados!';
-      _formKey = GlobalKey<FormState>();
-    });
-  }
-
-  void _calculate() {
-    setState(
-      () {
-        final weight = double.parse(weightController.text);
-        final height = double.parse(heightController.text) / 100;
-        final imc = weight / (height * height);
-        if (imc < 18.6) {
-          _infoText = 'Abaixo do Peso (${imc.toStringAsPrecision(4)}kg)';
-        } else if (imc >= 18.6 && imc < 24.9) {
-          _infoText = 'Peso Ideal (${imc.toStringAsPrecision(4)}kg)';
-        } else if (imc >= 24.9 && imc < 29.9) {
-          _infoText =
-              'Levemente Acima do Peso (${imc.toStringAsPrecision(4)}kg)';
-        } else if (imc >= 29.9 && imc < 34.9) {
-          _infoText = 'Obesidade Grau I (${imc.toStringAsPrecision(4)}kg)';
-        } else if (imc >= 34.9 && imc < 39.9) {
-          _infoText = 'Obesidade Grau II (${imc.toStringAsPrecision(4)}kg)';
-        } else if (imc >= 40) {
-          _infoText = 'Obesidade Grau III (${imc.toStringAsPrecision(4)}kg)';
-        }
-      },
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          'Calculadora de IMC',
-        ),
-        centerTitle: true,
-        backgroundColor: Colors.green,
-        actions: <Widget>[
-          IconButton(
-            icon: Icon(
-              Icons.refresh,
-            ),
-            onPressed: _resetFields,
-          )
-        ],
-      ),
-      backgroundColor: Colors.white,
-      body: SingleChildScrollView(
-        padding: EdgeInsets.fromLTRB(10, 0, 10, 0),
-        child: Form(
-          key: _formKey,
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.stretch,
-            children: <Widget>[
-              Icon(
-                Icons.person_outline,
-                size: 120,
-                color: Colors.green,
-              ),
-              TextFormField(
-                keyboardType: TextInputType.number,
-                decoration: InputDecoration(
-                  labelText: 'Peso (kg)',
-                  labelStyle: TextStyle(
-                    color: Colors.green,
-                  ),
-                ),
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.green,
-                  fontSize: 25,
-                ),
-                controller: weightController,
-                validator: (value) {
-                  if (value.isEmpty) {
-                    return 'Insira seu Peso!';
-                  }
-                  return value;
-                },
-              ),
-              TextFormField(
-                keyboardType: TextInputType.number,
-                decoration: InputDecoration(
-                  labelText: 'Altura (cm)',
-                  labelStyle: TextStyle(
-                    color: Colors.green,
-                  ),
-                ),
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.green,
-                  fontSize: 25,
-                ),
-                controller: heightController,
-                validator: (value) {
-                  if (value.isEmpty) {
-                    return 'Insira sua Altura!';
-                  }
-                  return value;
-                },
-              ),
-              Padding(
-                padding: EdgeInsets.only(top: 10, bottom: 10),
-                child: Container(
-                  height: 50,
-                  child: ElevatedButton(
-                    onPressed: _calculate,
-                    style: ElevatedButton.styleFrom(
-                      primary: Colors.green,
-                    ),
-                    child: Text(
-                      'Calcular',
-                      style: TextStyle(
-                        color: Colors.white,
-                        fontSize: 25,
-                      ),
-                    ),
-                  ),
-                ),
-              ),
-              Text(
-                _infoText,
-                textAlign: TextAlign.center,
-                style: TextStyle(
-                  color: Colors.green,
-                  fontSize: 25,
-                ),
-              )
-            ],
-          ),
-        ),
-      ),
-    );
-  }
 }
